A Simple Weather App using React.
A Simple Weather App using React. We get the current Weather data using API from OpenWeatherMap and display the weather information like Temperature,  Weather condition, Humidity and Wind Speed according to the city in this Weather app. In this Weather app, You just need to enter the name of the city and it gets you the accurate data using free API.